requirements.txt
kubernetes==12.0.1
pyyaml==5.3.1
